requirements.txt
requests
python-dotenv
markdown
weasyprint
caldav
icalendar
recurring-ical-events
pytz
